requirements.txt
black==25.1.0
flake8==7.1.1
Django==5.1.6
django-auditlog==3.0.0
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-filter==24.3
django-redis==5.4.0
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-extra-fields==3.7.0
drf-spectacular==0.28.0
pillow==11.1.0
psycopg2-binary==2.9.10
requests==2.32.3
